services/order-service/requirements.txt
flask==3.0.0
gunicorn==21.2.0
boto3==1.34.0
psycopg2-binary==2.9.9
redis==5.0.1
psutil==5.9.6
